Implemented custom ROS messages on Arduino. No more CSV parsing.
Note these will not work for you until I merge my catkin workspace into git (to be done) because of some dependences I needed to create.

Also note: the Arduino Uno cannot handle both the gyro and the other sensors because it doesn't have enough memory to store the ROS buffers, so we'll need the mega.